NODE-733 Support V2 for transactions in /sign and /broadcast handlers
